sorted attribute reexports
