+++ src/Iced/Style.hs
@@ -1,5 +1,30 @@
 module Iced.Style (
-  module Iced.Style.Container,
+  module Iced.Style.Internal,
+  module Iced.Style.Status,
 ) where
 
-import Iced.Style.Container
+--
+-- Reexport style attributes
+--
+
+import Iced.Style.Internal (
+  background,
+  bar,
+  border,
+  borderColor,
+  borderWidth,
+  dotColor,
+  handleColor,
+  iconColor,
+  placeholderColor,
+  selectionColor,
+  textColor,
+ )
+import Iced.Style.Status (
+  active,
+  disabled,
+  focused,
+  hovered,
+  opened,
+  pressed,
+ )
