--- src/Iced/Style/Container.hs
@@ -1,1 +1,0 @@
-module Iced.Style.Container where
