+++ src/Iced/Widget.hs
@@ -1,85 +1,59 @@
 module Iced.Widget (
-  module Iced.Widget.Button,
-  module Iced.Widget.Canvas,
-  module Iced.Widget.Checkbox,
-  module Iced.Widget.Column,
-  module Iced.Widget.ComboBox,
-  module Iced.Widget.Container,
-  module Iced.Widget.Image,
-  module Iced.Widget.Markdown,
-  module Iced.Widget.MouseArea,
-  module Iced.Widget.PickList,
-  module Iced.Widget.ProgressBar,
-  module Iced.Widget.Radio,
-  module Iced.Widget.Responsive,
-  module Iced.Widget.Row,
-  module Iced.Widget.Scrollable,
-  module Iced.Widget.Slider,
+  button,
+  canvas,
+  checkbox,
+  column,
+  comboBox,
+  container,
+  image,
+  markdown,
+  mouseArea,
+  pickList,
+  progressBar,
+  radio,
+  responsive,
+  row,
+  scrollable,
+  slider,
   module Iced.Widget.Space,
-  module Iced.Widget.Text,
-  module Iced.Widget.TextEditor,
-  module Iced.Widget.TextInput,
-  module Iced.Widget.Toggler,
-  module Iced.Widget.Tooltip,
-  module Iced.Widget.VerticalSlider,
+  text,
+  textEditor,
+  textInput,
+  toggler,
+  tooltip,
+  verticalSlider,
 ) where
 
 --
--- Reexport only key widget constructors and attributes
+-- Reexport only key widget constructors
 --
 
-import Iced.Widget.Button (
-  button,
-  onPressIf,
- )
+import Iced.Widget.Button (button)
 import Iced.Widget.Canvas (canvas)
-import Iced.Widget.Checkbox (
-  checkbox,
-  onToggle,
-  onToggleIf,
-  textLineHeight,
-  textShaping,
-  textSize,
- )
-import Iced.Widget.Column
-import Iced.Widget.ComboBox hiding (State, newState)
-import Iced.Widget.Container (
-  centerX,
-  centerY,
-  container,
- )
+import Iced.Widget.Checkbox (checkbox)
+import Iced.Widget.Column (column)
+import Iced.Widget.ComboBox (comboBox)
+import Iced.Widget.Container (container)
 import Iced.Widget.Image (image)
 import Iced.Widget.Markdown (markdown)
-import Iced.Widget.MouseArea (
-  mouseArea,
-  onDoubleClick,
-  onEnter,
-  onExit,
-  onMiddlePress,
-  onMiddleRelease,
-  onMove,
-  onRelease,
-  onRightPress,
-  onRightRelease,
- )
+import Iced.Widget.MouseArea (mouseArea)
 import Iced.Widget.PickList (pickList)
 import Iced.Widget.ProgressBar (progressBar)
 import Iced.Widget.Radio (radio)
-import Iced.Widget.Responsive
-import Iced.Widget.Row
-import Iced.Widget.Scrollable
-import Iced.Widget.Slider
-import Iced.Widget.Space
+import Iced.Widget.Responsive (responsive)
+import Iced.Widget.Row (row)
+import Iced.Widget.Scrollable (scrollable)
+import Iced.Widget.Slider (slider)
+import Iced.Widget.Space (
+  horizontalSpace,
+  space,
+  spaceHeight,
+  spaceWidth,
+  verticalSpace,
+ )
 import Iced.Widget.Text (text)
-import Iced.Widget.TextEditor (
-  onAction,
-  textEditor,
- )
-import Iced.Widget.TextInput (
-  onInput,
-  onSubmit,
-  textInput,
- )
-import Iced.Widget.Toggler
-import Iced.Widget.Tooltip hiding (Position (..))
-import Iced.Widget.VerticalSlider
+import Iced.Widget.TextEditor (textEditor)
+import Iced.Widget.TextInput (textInput)
+import Iced.Widget.Toggler (toggler)
+import Iced.Widget.Tooltip (tooltip)
+import Iced.Widget.VerticalSlider (verticalSlider)
