+++ src/Iced/Widget/Container.hs
@@ -19,6 +19,7 @@ import Iced.Attribute.PaddingFFI
 import Iced.Attribute.Style
 import Iced.ColorFFI
 import Iced.Element
+import Iced.Style.Internal
 import Iced.Theme
 
 data NativeContainer
