+++ src/Iced/Widget/PickList.hs
@@ -19,10 +19,11 @@ import Iced.Attribute.Internal
 import Iced.Attribute.LengthFFI
 import Iced.Attribute.PaddingFFI
 import Iced.Attribute.Placeholder
-import Iced.Attribute.Status
 import Iced.Attribute.Style
 import Iced.ColorFFI
 import Iced.Element
+import Iced.Style.Internal
+import Iced.Style.Status
 import Iced.Theme
 
 data NativePickList
