+++ src/Iced/Widget/TextInput.hs
@@ -21,10 +21,11 @@ import Iced.Attribute.Internal
 import Iced.Attribute.LengthFFI
 import Iced.Attribute.OnInput
 import Iced.Attribute.PaddingFFI
-import Iced.Attribute.Status
 import Iced.Attribute.Style
 import Iced.ColorFFI
 import Iced.Element
+import Iced.Style.Internal
+import Iced.Style.Status
 import Iced.Theme
 
 data NativeTextInput
